requirements.txt
discord.py
pytest
logzero